@@ -1,9 +1,11 @@
 package top.sharehome.springbootadmin.config;
 
 import de.codecentric.boot.admin.server.config.AdminServerProperties;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.SavedRequestAwareAuthenticationSuccessHandler;
 
 /**
@@ -12,27 +14,34 @@ import org.springframework.security.web.authentication.SavedRequestAwareAuthenti
  * @author AntonyCheng
  */
 @Configuration(proxyBeanMethods = false)
-public class SecuritySecureConfig extends WebSecurityConfigurerAdapter {
+public class SecuritySecureConfig {
+
     private final String adminContextPath;
 
     public SecuritySecureConfig(AdminServerProperties adminServerProperties) {
         this.adminContextPath = adminServerProperties.getContextPath();
     }
 
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
         SavedRequestAwareAuthenticationSuccessHandler successHandler = new SavedRequestAwareAuthenticationSuccessHandler();
         successHandler.setTargetUrlParameter("redirectTo");
 
-        http.authorizeRequests()
-                .antMatchers(adminContextPath + "/assets/**").permitAll()
-                .antMatchers(adminContextPath + "/login").permitAll()
-                .anyRequest().authenticated()
-                .and()
-                .formLogin().loginPage(adminContextPath + "/login").successHandler(successHandler).and()
-                .logout().logoutUrl(adminContextPath + "/logout").and()
-                .httpBasic().and()
-                .csrf().disable();
-    }
+        http
+                .authorizeHttpRequests(auth -> auth
+                        .requestMatchers(adminContextPath + "/assets/**").permitAll()
+                        .requestMatchers(adminContextPath + "/login").permitAll()
+                        .anyRequest().authenticated()
+                )
+                .formLogin(form -> form
+                        .loginPage(adminContextPath + "/login")
+                        .successHandler(successHandler)
+                )
+                .logout(logout -> logout.logoutUrl(adminContextPath + "/logout"))
+                .httpBasic(httpBasic -> {
+                })
+                .csrf(AbstractHttpConfigurer::disable);
 
+        return http.build();
+    }
 }
